Add code to make dependency lookup a bit smarter
I think that sometimes it was failing because, depending on where you
were running it from and the current directory, sometimes the paths in
appsettings didn't get interpreted correctly and it would look in the
wrong place. I'm not totally sure that was the problem but moving the
dependency lookup code into its own place is a good change anyway I
think.